% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{assert_fow}
\alias{assert_fow}
\title{Run checks on a taxonomic database}
\usage{
assert_fow(fow)
}
\arguments{
\item{fow}{Dataframe; taxonomic database in Darwin Core format}
}
\value{
Dataframe; taxonomic database in Darwin Core format. Will
be the same as the input.
}
\description{
Stops if any check fails.
}
