% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dct_filmies}
\alias{dct_filmies}
\title{Taxonomic data of filmy ferns}
\format{
Dataframe (tibble), with 2729 rows and
5 columns. For details about data format, see
\url{https://dwc.tdwg.org/terms/#taxon}.
}
\source{
\url{https://dwc.tdwg.org/terms/#taxon}
}
\usage{
dct_filmies
}
\description{
Taxonomic data of filmy ferns (family Hymenophyllaceae) in Darwin Core format.
Modified from data downloaded from the \href{https://www.catalogueoflife.org/}{Catalog of Life} under the \href{https://creativecommons.org/licenses/by/4.0/}{Creative Commons Attribution (CC BY) 4.0}
license.
}
\keyword{datasets}
