% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dct_fix_format.R
\name{dct_fix_format}
\alias{dct_fix_format}
\title{Fix the format of a taxonomic database to conform with Darwin Core Taxon standards}
\usage{
dct_fix_format(tax_dat)
}
\arguments{
\item{tax_dat}{Dataframe; taxonomic database that is supposed to be in Darwin
Core format}
}
\value{
Dataframe; taxonomic database in Darwin Core format
}
\description{
Fix the format of a taxonomic database to conform with Darwin Core Taxon standards
}
