% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dct_terms}
\alias{dct_terms}
\title{Darwin Core Taxon terms}
\format{
Character vector, including two attributes:
\itemize{
\item \code{retrieved}: Date the terms were obtained
\item \code{url}: URL from which the terms were obtained
}
}
\source{
\url{https://dwc.tdwg.org/terms/#taxon}
}
\usage{
dct_terms
}
\description{
A list of Darwin Core terms in the Taxon class
}
\keyword{datasets}
