% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dwctaxon-package.R
\docType{package}
\name{dwctaxon-package}
\alias{dwctaxon}
\alias{dwctaxon-package}
\title{dwctaxon: Tools for Working with Darwin Core Taxon Data}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Provides functions to create, manipulate, and validate taxonomic data in compliance with Darwin Core standards (Darwin Core "Taxon" class https://dwc.tdwg.org/terms/#taxon).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://docs.ropensci.org/dwctaxon/}
  \item \url{https://github.com/ropensci/dwctaxon}
  \item Report bugs at \url{https://github.com/ropensci/dwctaxon/issues}
}

}
\author{
\strong{Maintainer}: Joel H. Nitta \email{joelnitta@gmail.com} (\href{https://orcid.org/0000-0003-4719-7472}{ORCID})

Other contributors:
\itemize{
  \item Wataru Iwasaki (\href{https://orcid.org/0000-0002-9169-9245}{ORCID}) [contributor]
  \item Collin Schwantes (Collin reviewed the package (v. 1.0.0.9000) for rOpenSci, see <https://github.com/ropensci/software-review/issues/574>) [reviewer]
  \item Stephen Formel (Stephen reviewed the package (v. 1.0.0.9000) for rOpenSci, see <https://github.com/ropensci/software-review/issues/574>) [reviewer]
}

}
\keyword{internal}
