% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcg_assert_tax_dat.R
\name{pcg_assert_tax_dat}
\alias{pcg_assert_tax_dat}
\title{Assert that a taxonomic database is correctly formatted}
\usage{
pcg_assert_tax_dat(tax_dat)
}
\arguments{
\item{tax_dat}{Dataframe; taxonomic database in Darwin Core format}
}
\value{
Dataframe; taxonomic database in Darwin Core format. Will
be the same as the input.
}
\description{
Assertions include:
\itemize{
\item all synonyms must map to an accepted name
\item all synonyms and accepted names must be distinct
\item all names must include a unique taxonID
}
}
\details{
Stops with an error if any check fails.
}
