% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcg_extract_fow.R
\name{pcg_extract_fow}
\alias{pcg_extract_fow}
\title{Extract Ferns of the World taxonomic data from Catalog of Life}
\usage{
pcg_extract_fow(col_data)
}
\arguments{
\item{col_data}{Catalog of Life data. Output of of load_col()}
}
\value{
Dataframe (tibble)
}
\description{
Also do some quality checks, and filter to only data at species level or below
}
