% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcg_load_col.R
\name{pcg_load_col}
\alias{pcg_load_col}
\title{Load data from the Catalog of Life}
\usage{
pcg_load_col(col_data_path)
}
\arguments{
\item{col_data_path}{Path to data downloaded from the Catalog of Life; either
the original zip file, or the tsv file containing the taxonomic data extracted
from the zip file (usually named "Taxon.tsv").}
}
\value{
Dataframe (tibble)
}
\description{
Catalog of Life data can be downloaded from here:
https://download.catalogueoflife.org/col/monthly/. The "_dwca" format should
be selected when downloading (files ending in "_dwca.zip").
}
\details{
If \code{col_data_path} ends in "zip", the function will try to automatically
extract the taxonomic names data, which is typically named "Taxon.tsv".
}
