% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcg_add_row.R
\name{dct_add_row}
\alias{dct_add_row}
\title{Add one entry to a taxonomic database}
\usage{
dct_add_row(
  tax_dat,
  taxon_id = NULL,
  sci_name,
  tax_status,
  usage_id = NULL,
  strict = FALSE
)
}
\arguments{
\item{tax_dat}{Dataframe; taxonomic database in Darwin Core format}

\item{taxon_id}{taxonID to use for new entry; optional, will be
assigned automatically if not provided.}

\item{sci_name}{scientificName to use for new entry}

\item{tax_status}{taxonomicStatus to use for new entry}

\item{usage_id}{acceptedNameUsageID to use for new entry; optional}

\item{strict}{Logical; should taxonomic checks be run on the updated
taxonomic database?}
}
\value{
Dataframe; taxonomic database in Darwin Core format
}
\description{
Add one entry to a taxonomic database
}
