% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcg_change_status.R
\name{dct_change_status}
\alias{dct_change_status}
\title{Change the taxonomic status of one entry in a taxonomic database}
\usage{
dct_change_status(
  tax_dat,
  taxon_id = NULL,
  sci_name = NULL,
  new_status,
  strict = FALSE
)
}
\arguments{
\item{tax_dat}{Dataframe; taxonomic database in Darwin Core format}

\item{taxon_id}{taxonID for the entry to be changed}

\item{sci_name}{scientificName for the entry to be changed}

\item{new_status}{Updated taxonomicStatus to use for the entry}

\item{strict}{Logical; should taxonomic checks be run on the updated
taxonomic database?}
}
\value{
Dataframe; taxonomic database in Darwin Core format
}
\description{
Only one of \code{taxon_id} or \code{sci_name} needs to be entered. Either will
work if as long as it makes a partial or full match to one row in the data.
}
