% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dct_check_mapping.R
\name{dct_check_mapping}
\alias{dct_check_mapping}
\title{Check mapping of usage names}
\usage{
dct_check_mapping(tax_dat, on_fail, on_success)
}
\arguments{
\item{tax_dat}{Dataframe; taxonomic database in DWC format.}

\item{on_fail}{Character vector of length 1, either "error" or "summary". Describes what to do if the check fails. Default \code{"error"}.}

\item{on_success}{Character vector of length 1, either "logical" or "data". Describes what to do if the check passes. Default \code{"data"}.}
}
\value{
Depends on the result of the check and on values of \code{on_fail} and
\code{on_success}:
\itemize{
\item If the check passes and \code{on_success} is "logical", return \code{TRUE}
\item If the check passes and \code{on_success} is "data", return the input dataframe
\item If the check fails and \code{on_fail} is "error", return an error
\item If the check fails and \code{on_fail} is "summary", issue a warning and
return a dataframe with a summary of the reasons for failure
}
}
\description{
Check that names with acceptedUsageID map properly to taxonID in
Darwin Core taxonomic data
}
\details{
The following rules are enforced:
\itemize{
\item taxonID may not be identical to acceptedNameUsageID within a single row
\item Every acceptedNameUsageID must have a corresponding taxonID
}
}
\examples{
# The bad data has an acceptedNameUsageID (third row, "4") that lacks a
# corresponding taxonID
bad_dat <- tibble::tribble(
  ~taxonID, ~acceptedNameUsageID, ~taxonomicStatus, ~scientificName,
  "1", NA, "accepted", "Species foo",
  "2", "1", "synonym", "Species bar",
  "3", "4", "synonym", "Species bat"
)
suppressWarnings(
  dct_check_mapping(bad_dat, on_fail = "summary")
)
}
