% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dct_validate.R
\name{dct_validate}
\alias{dct_validate}
\title{Check that a taxonomic database is correctly formatted}
\usage{
dct_validate(
  tax_dat,
  check_taxon_id = TRUE,
  check_mapping = TRUE,
  check_taxonomic_status = TRUE,
  check_acc_syn_diff = TRUE,
  check_col_names = TRUE
)
}
\arguments{
\item{tax_dat}{Dataframe; taxonomic database in Darwin Core format}

\item{check_taxon_id}{Logical; should all instances of \code{taxonID} be required
to be non-missing and unique?}

\item{check_mapping}{Logical; should all values of \code{acceptedNameUsageID} be
required to map to the \code{taxonID} of an existing name?}

\item{check_taxonomic_status}{Logical; should all taxonomic names be required
to include the status of either "accepted" or "synonym"?}

\item{check_acc_syn_diff}{Logical; should accepted names and synonyms be
required to be different?}

\item{check_col_names}{Logical; should all column names be required to
be a valid Darwin Core term?}
}
\value{
Dataframe; taxonomic database in Darwin Core format. Will
be the same as the input.
}
\description{
Stops with an error if any check fails. Most checks are geared towards being
able to use the taxonomic database for taxonomic name resolution at the
species level.
}
\details{
For \code{check_taxonomic_status} and \code{check_acc_syn_diff}, "accepted" and
"synonym" are determined by string matching of \code{taxonomicStatus}; so
"provisionally accepted" is counted as "accepted", "ambiguous synonym" is
counted as "synonym", etc (not case-sensitive).
}
\examples{
dct_validate(dct_filmies)
}
