% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dct_validate.R
\name{dct_validate}
\alias{dct_validate}
\title{Check that a taxonomic database is correctly formatted}
\usage{
dct_validate(
  tax_dat,
  check_taxon_id = TRUE,
  check_mapping = TRUE,
  check_taxonomic_status = TRUE,
  check_acc_syn_diff = TRUE,
  check_col_names = TRUE,
  strict_mapping = TRUE,
  valid_tax_status = Sys.getenv("VALID_TAX_STATUS", unset =
    "accepted, synonym, variant, NA")
)
}
\arguments{
\item{tax_dat}{Dataframe; taxonomic database in Darwin Core format}

\item{check_taxon_id}{Logical; should all instances of \code{taxonID} be required
to be non-missing and unique?}

\item{check_mapping}{Logical; should all values of \code{acceptedNameUsageID} be
required to map to the \code{taxonID} of an existing name?}

\item{check_taxonomic_status}{Logical; should all taxonomic names be required
to include a valid value for taxonomic status (by default, "accepted",
"synonym", or "variant")?}

\item{check_acc_syn_diff}{Logical; should each scientific name be allowed
to have only one taxonomic status?}

\item{check_col_names}{Logical; should all column names be required to
be a valid Darwin Core term?}

\item{strict_mapping}{Logical; should rules about mapping of variants and
synonyms be enforced? (see Details)}

\item{valid_tax_status}{Character vector of length 1; valid values for
\code{taxonomicStatus}. Each value must be separated by a comma. Default
"accepted, synonym, variant, NA". "NA" indicates that missing (NA) values
are valid. Case-sensitive. Can also be set with the environmental variable
"VALID_TAX_STATUS".}
}
\value{
Dataframe; taxonomic database in Darwin Core format. Will
be the same as the input.
}
\description{
Stops with an error if any check fails. Most checks are geared towards being
able to use the taxonomic database for taxonomic name resolution at the
species level.
}
\details{
For \code{check_acc_syn_diff} and \code{strict_mapping}, "accepted", "synonym", and
"variant" are determined by string matching of \code{taxonomicStatus}; so
"provisionally accepted" is counted as "accepted", "ambiguous synonym" is
counted as "synonym", etc (case-sensitive).

For \code{strict_mapping}, the following rules are enforced:
\itemize{
\item Rows with \code{taxonomicStatus} of "synonym" (synonyms) must have an
\code{acceptedNameUsageID} matching the \code{taxonID} of an accepted name (
\code{taxonomicStatus} of "accepted")
\item Rows with \code{taxonomicStatus} of "variant" (orthographic variants) must
have an \code{acceptedNameUsageID} matching the \code{taxonID} of an accepted name or
synonym (but not another variant)
\item Rows with \code{taxonomicStatus} of "accepted" must not have any value entered
for \code{acceptedNameUsageID}
\item Rows with a value for \code{acceptedNameUsageID} must have a valid value for
\code{taxonomicStatus}.
}
}
\examples{

dct_validate(
  dct_filmies,
  valid_tax_status =
    paste("accepted name, ambiguous synonym, provisionally accepted name,",
          "synonym, variant")
)

}
